fixes merge conflict

+++ src/app/services/certificate-storage/certificate-storage.service.ts
@@ -10,12 +10,20 @@ export class CertificateStorageService {
 
   getCertificates(): ICertificate[] {
     const data = localStorage.getItem(this.storageKey);
+    this.certificates = data ? JSON.parse(data) : [];
 
-    return data ? JSON.parse(data) : [];
+    return this.certificates;
   }
 
   saveCertificates(certificate: ICertificate): void {
+    if (this.checkCertificateExists(certificate)) {
+      return;
+    }
     this.certificates.push(certificate);
     localStorage.setItem(this.storageKey, JSON.stringify(this.certificates));
+  }
+
+  checkCertificateExists(certificate: ICertificate): boolean {
+    return this.certificates.some(cert => cert.commonName === certificate.commonName);
   }
 }
